Fit factor models for ordinal data with pairwise likelihood methods

Fit latent variable models for ordinal variable using pairwise likelihood methods running via quasi-Newton BFGS or stochastic approximations.

Usage

fit_plFA(
  DATA,
  CONSTR_LIST,
  METHOD = c("ucminf", "SA"),
  INIT_METHOD = c("SA", "custom", "standard"),
  NCORES = 1,
  VALDATA = NULL,
  CONTROL = list(),
  CPP_CONTROL_MAIN = NULL,
  CPP_CONTROL_INIT = NULL,
  INIT = NULL,
  VERBOSE = FALSE
)

Arguments

DATA

Integer data matrix of dimension \(n*p\). Categories must be coded starting from zero.

CONSTR_LIST

List of constraints. It must contain

CONSTRMAT	\(p*q\)-dimensional matrix. Elements set to NA refers to free loading parameters. Elements set to numerical values denote fixed values constraints.

CONSTRVAR	\(q\)-dimensional vector. Elements set to NA refers to free latent variance parameters. Elements set to numerical values denote fixed values constraints.

CORRFLAG	Logical indicator. Set it to FALSE if the latent variables are assumed to be independent. Set it TRUE otherwise.

STDLV	Logical indicator. Set it to TRUE to fix latent variables scale. Set it FALSE to fix loadings scale.

LLC	Linear loadings constraints. Expects a list of constraints. See DETAILS.

METHOD

Label for the method chosen. Possible values are:

'ucminf'	for estimation via the quasi-Newton BFGS optimiser from the ucminf package. Used as default method.

'SA'	for estimation via Stochastic Approximations.

INIT_METHOD

'custom'	Uses the vector provided via INIT as starting point.

'standard'	Uses cold initialisation where loadings are initialised at 0.5 and the latent covariance matrix as an Identity matrix.

'SA'	Computes the starting point using a short chain of stochastic updates. Updates start from the standard initialisation point. Set as default method.

NCORES

Integer value setting the number of threads to carry out the estimation.

VALDATA

Validation data used to monitor convergence of stochastic approximations. If NULL, data passed via DATA is used for monitoring purposes.

CONTROL

List of control options to pass to ucminf. See ucminf::ucminf documentation.

CPP_CONTROL_MAIN

List of control options to pass to the SA optimiser when METHOD="SA":

PAIRS_PER_ITERATION	Number of pairs to draw at each iteration.

MAXT	Maximum number of iterations.

BURN	Maximum number of iterations to burn before trajectory averaging.

STEP0	Initial step length.

CPP_CONTROL_INIT

List of control options to pass to the SA optimiser when INIT_METHOD="SA". See CPP_CONTROL_MAIN for details.

INIT

Initialising vector. If not provided, the starting point is computed according to INIT_METHOD

VERBOSE

TRUE for verbose output

Value

Object of class plFaFit.

Details

The argument CONSTR_LIST$LLC is expected to be a list of constraints, e.g. CONSTR_LIST$LLC <- list(constraint_1, constraint2, ...), where each constraint is defined by a list itself. For example, to impose the constraint "L_(2,1)=0.5L_(5,2)+0.25L_(9,3)" you have to write constraint_1 <- list(c(2,1), c(0.5,5,2), c(0.25, 9,3)). That is: the first vector of the list is 2-dimensional and stores the coordinates of the constrained loading. Each of the successive triplets represent a linear coefficient followed by the coordinates of the loading of reference. You can set up an arbitrary number of triplets in each constraint.